requirements.txt
streamlit
torch
torchvision
Pillow
rembg
trimesh
numpy
transformers
einops
omegaconf
timm
imageio
imageio-ffmpeg
